get rid of `MultiProviderConnector` and move the logic to the datamodel_connector crate
